Updated NavBar, LandingPage, App.js and index.html
Added fonts library to the index.html file. Updated App to include the theme and font-style for all existing elements. Removed the shaped and updated with a background image on the landing page. Fixed some NavBar items. Updated the landing_image.png file to remove background when it opens to make the background visible,
Other minor code and design changes.