Updates to game config
Modified version of the code to make it so that when the game.revealAnswers property is false, it just gives user a generic message to the user.
It now only displays Congratulations or You are incorrect (along with the words) when game.revealAnswers is true.

Also, if session is timed and the multiple submissions are not allowed (i.e. allowMultipleSubmissions = false) after a user submits the results and closes the confirmation pop-up, the user leaves the session automatically now.
If the game.multipleSubmissions are allowed but the timer runs out, the answers now get auto submitted (and the follow the logic mentioned above for game.revealAnswers) and then once users closes that dialog box, the leaveSession() function is be triggered.